SPARK-51252 Adding state store level metrics for last uploaded snapshot version in HDFS State Stores
